[RHTAP-1303] Divided methods to files + comments
